+++ common/community.php
@@ -1,0 +1,29 @@
+<!-- Community Section -->
+<div id="community" class="text-center">
+  <div class="container">
+    <div class="section-title center">
+      <h2>COMMUNITY</h2>
+      <hr>
+    </div>
+    <div class="row col-md-8 col-md-offset-2">
+      <div class="col-md-6 col-sm-6 wow fadeInDown" data-wow-delay="400ms">
+        <a href="https://discordapp.com/invite/K5gPZ2h" target="_blank">
+          <div class="community-box">
+            <i class="fab fa-discord fa-5x"></i>
+            <span class="count">2016</span>
+            <h4>Members</h4>
+          </div>
+        </a>
+      </div>
+      <div class="col-md-6 col-sm-6 wow fadeInDown" data-wow-delay="200ms">
+        <a href="https://twitter.com/VIPSTARTOKEN" target="_blank">
+          <div class="community-box">
+            <i class="fab fa-twitter fa-5x"></i>
+            <span class="count">1882</span>
+            <h4>Followers</h4>
+          </div>
+        </a>
+      </div>
+    </div>
+  </div>
+</div>
