+++ common/contact.php
@@ -1,0 +1,34 @@
+<!-- Contact Section -->
+<div id="contact" class="text-center">
+  <div class="container">
+    <div class="section-title center">
+      <h2>CONTACT</h2>
+      <hr>
+    </div>
+    <div class="col-md-8 col-md-offset-2">
+      <div class="col-md-6">
+        <a href="mailto:vipstarcoin&#64;gmail.com">
+          <i class="far fa-envelope fa-5x"></i>
+          <p>vipstarcoin&#64;gmail.com</p>
+        </a>
+      </div>
+      <div class="col-md-6">
+        <a href="https://twitter.com/VIPSTARTOKEN" target="_blank">
+          <i class="fab fa-twitter fa-5x"></i>
+          <p>@VIPSTARTOKEN</p>
+        </a>
+      </div>
+      <div class="clearfix"></div>
+      <div class="contact-tweet">
+        <a class="twitter-timeline"
+           data-width="500"
+           data-height="500"
+           data-theme="light"
+           href="https://twitter.com/VIPSTARTOKEN?ref_src=twsrc%5Etfw">
+          Tweets by VIPSTARTOKEN
+        </a>
+        <script async src="https://platform.twitter.com/widgets.js" charset="utf-8"></script>
+      </div>
+    </div>
+  </div>
+</div>
