+++ common/head.php
@@ -1,0 +1,43 @@
+<!-- Global site tag (gtag.js) - Google Analytics -->
+<script async src="https://www.googletagmanager.com/gtag/js?id=UA-117099829-1"></script>
+<script>
+  window.dataLayer = window.dataLayer || [];
+  function gtag(){dataLayer.push(arguments);}
+  gtag('js', new Date());
+
+  gtag('config', 'UA-117099829-1');
+</script>
+
+<meta charset="utf-8">
+<meta name="viewport" content="width=device-width, initial-scale=1">
+<title>VIPSTARCOIN</title>
+
+<meta property="og:title" content="VIPSTARCOIN">
+<meta property="og:type" content="website">
+<meta property="og:url" content="<?php echo URL; ?>/">
+<meta property="og:image" content="<?php echo URL; ?>/img/og-image.jpg">
+<meta property="og:site_name" content="VIPSTARCOIN">
+<meta name="twitter:card" content="summary_large_image">
+
+<!-- Favicons -->
+<link rel="shortcut icon" href="<?php echo URL; ?>/img/favicon.png" type="image/x-icon">
+<link rel="apple-touch-icon" href="<?php echo URL; ?>/img/apple-touch-icon.png">
+
+<!-- Bootstrap -->
+<link rel="stylesheet"
+      href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/css/bootstrap.min.css"
+      integrity="sha384-BVYiiSIFeK1dGmJRAkycuHAHRg32OmUcww7on3RYdg4Va+PmSTsz/K68vbdEjh4u"
+      crossorigin="anonymous">
+<link href="https://use.fontawesome.com/releases/v5.0.8/css/all.css" rel="stylesheet">
+
+<!-- Stylesheet -->
+<link rel="stylesheet" type="text/css"  href="<?php echo URL; ?>/css/style.css">
+<link rel="stylesheet" type="text/css" href="<?php echo URL; ?>/css/prettyPhoto.css">
+<link href='http://fonts.googleapis.com/css?family=Lato:400,700,900,300' rel='stylesheet' type='text/css'>
+<link href='http://fonts.googleapis.com/css?family=Open+Sans:400,700,800,600,300' rel='stylesheet' type='text/css'>
+<!-- HTML5 shim and Respond.js for IE8 support of HTML5 elements and media queries -->
+<!-- WARNING: Respond.js doesn't work if you view the page via file:// -->
+<!--[if lt IE 9]>
+  <script src="https://oss.maxcdn.com/html5shiv/3.7.2/html5shiv.min.js"></script>
+  <script src="https://oss.maxcdn.com/respond/1.4.2/respond.min.js"></script>
+<![endif]-->
