+++ common/header.php
@@ -1,0 +1,4 @@
+<!-- Header -->
+<header id="header" class="header">
+  <img src="<?php echo URL; ?>/img/header.jpg" alt="VIPSTAR" class="header-image">
+</header>
