+++ common/nav.php
@@ -1,0 +1,28 @@
+<!-- Navigation -->
+<div id="nav">
+  <nav class="navbar navbar-custom">
+    <div class="container">
+      <div class="navbar-header">
+        <button type="button" class="navbar-toggle" data-toggle="collapse" data-target=".navbar-main-collapse">
+          <i class="fa fa-bars"></i>
+        </button>
+        <a class="navbar-brand page-scroll" href="#page-top">VIPSTARCOIN</a>
+      </div>
+
+      <!-- Collect the nav links, forms, and other content for toggling -->
+      <div class="collapse navbar-collapse navbar-right navbar-main-collapse">
+        <ul class="nav navbar-nav">
+          <!-- Hidden li included to remove active class from about link when scrolled up past about section -->
+          <li class="hidden"> <a href="#page-top"></a> </li>
+          <li> <a class="page-scroll" href="#about">ABOUT</a> </li>
+          <li> <a class="page-scroll" href="#roadmap">ROADMAP</a> </li>
+          <li> <a class="page-scroll" href="#holding">HOLDING</a> </li>
+          <li> <a class="page-scroll" href="#wallet">WALLET</a> </li>
+          <li> <a class="page-scroll" href="#news">NEWS</a> </li>
+          <li> <a class="page-scroll" href="#community">COMMUNITY</a> </li>
+          <li> <a class="page-scroll" href="#contact">CONTACT</a> </li>
+        </ul>
+      </div>
+    </div>
+  </nav>
+</div>
