+++ common/script.php
@@ -1,0 +1,15 @@
+<script src="https://code.jquery.com/jquery-2.2.4.min.js"
+        integrity="sha256-BbhdlvQf/xTY9gja0Dq3HiwQF8LaCRTXxZKRutelT44="
+        crossorigin="anonymous"></script>
+<script src="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/js/bootstrap.min.js"
+        integrity="sha384-Tc5IQib027qvyjSMfHjOMaLkfuWVxZxUPnCJA7l2mCWNIpG9mGCD8wGNIcPD7Txa"
+        crossorigin="anonymous"></script>
+<script src="https://cdnjs.cloudflare.com/ajax/libs/Chart.js/2.7.2/Chart.bundle.min.js"
+        integrity="sha256-XF29CBwU1MWLaGEnsELogU6Y6rcc5nCkhhx89nFMIDQ="
+        crossorigin="anonymous"></script>
+<script type="text/javascript" src="<?php echo URL; ?>/js/jquery.isotope.js"></script>
+<script type="text/javascript" src="<?php echo URL; ?>/js/jquery.counterup.js"></script>
+<script type="text/javascript" src="<?php echo URL; ?>/js/waypoints.js"></script>
+<script type="text/javascript" src="<?php echo URL; ?>/js/jquery.inview.min.js"></script>
+<script type="text/javascript" src="<?php echo URL; ?>/js/main.js"></script>
+<script type="text/javascript" src="//s7.addthis.com/js/300/addthis_widget.js#pubid=ra-5ab60eab9e6f635f"></script>
