+++ en/about.php
@@ -1,0 +1,137 @@
+<!-- About Section -->
+<div id="about">
+  <div class="container">
+    <div class="dropdown">
+      <button class="button button-bw dropdown-toggle" type="button" id="dropdownMenu1" data-toggle="dropdown" aria-haspopup="true" aria-expanded="true">
+        <i class="fas fa-language"></i>LANGUAGE
+        <span class="caret"></span>
+      </button>
+      <ul class="dropdown-menu" aria-labelledby="dropdownMenu1">
+        <li class="selected">English<i class="fas fa-check"></i></li>
+        <li role="separator" class="divider"></li>
+        <li><a href="<?php echo URL; ?>/">日本語</a></li>
+        <li><a href="<?php echo URL; ?>/vip/">VIP</a></li>
+      </ul>
+    </div>
+
+    <div class="section-title text-center center">
+      <h2>ABOUT</h2>
+      <hr>
+    </div>
+    <div class="row">
+      <div class="col-md-12 text-center"><img src="<?php echo URL; ?>/img/about.jpg" class="js_about-img img-responsive"></div>
+      <div class="col-md-8 col-md-offset-2">
+        <div class="about-text">
+        <p>VIPSTARCOIN (VIPS) is a cryptocurrency originated from a community called 'VIP' in 5chnnel, the largest Japanese textboards.</p>
+        <p>VIPS is currently being developed by a volunteer programmer and test users. We dev team is targeting for VIPS to be utilized in varied ways as an original cryptocurrency from Japan, setting the final goal for contributing to the society across borders as a convenient method of making donations.</p>
+        <p>We hope VIPS gains popularity in the real society as well as on the Internet community with the strong support of VIPS holders.</p>
+          <p><a href="#" class="about-qa" data-toggle="modal" data-target="#qaModal">【FOR BEGINNERS】What is a cryptocurrency? Q＆A&nbsp;&nbsp;<i class="far fa-window-restore"></i></a></p>
+          <p class="bl_about-button text-center">
+            <a class="button button-bw" href="#" data-toggle="modal" data-target="#pointModal">6 unique features of VIPS&nbsp;&nbsp;<i class="far fa-window-restore"></i></a>
+          </p>
+          <!-- Modal -->
+          <div class="modal fade" id="qaModal" tabindex="-1" role="dialog" aria-labelledby="qaModalLabel">
+            <div class="modal-dialog" role="document">
+              <div class="modal-content">
+                <div class="modal-header">
+                  <button type="button" class="close" data-dismiss="modal" aria-label="Close"><span aria-hidden="true">×</span></button>
+                  <h4 class="modal-title" id="qaModalLabel">What is a cryptocurrency? Q＆A</h4>
+                </div>
+                <div class="modal-body">
+                <p>
+                  <span>Q. What is the 'VIP' community?</span><br>
+                  <span><b>A. The 'VIP' community derives from an online textboard titled 'NyuSoku VIP(ニュー速VIP)' on 5ch.net. It is the second active community of 5ch following NANJ (Nandemo Jikkyo, なんでも実況). Community members of 'VIP' are called as 'VIPPER.'</b></span>
+                </p>
+                <p>
+                  <span>Q. What is VIPSTARCOIN (VIPS)? </span><br>
+                  <span><b>A. VIPS is a cryptocurrency that is originated from the VIP community. VIPS is created by an anonymous vipper and is equipped with cutting-edge technologies such as smart contracts and Proof of Stake version 3. One can create new tokens from VIPS just as they can do so with Ethereum.</b></span>
+                </p>
+                <p>
+                  <span>Q. What is a cryptocurrency？</span><br>
+                  <span><b>A. A cryptocurrency is a series of blocks of data created based on the blockchain technology and is hard to be copied or tampered.</b></span>
+                </p>
+                <p>
+                  <span>Q. How can I receive or keep a cryptocurrency?</span><br>
+                  <span><b>A. You need a virtual wallet to send or receive cryptocurrency coins or tokens. Different types of wallets might be needed to keep multiple kinds of cryptocurrencies. For VIPS, you need a special wallet arranged for VIPS coins. Ethereum wallets such as MyEtherWallet(MEW) cannot be used for VIPS coins or tokens.</b></span>
+                </p>
+                <p>
+                  <span>Q. Do I need to provide personal information to the developer for making a wallet for VIPS coins?</span><br>
+                  <span><b>A. No, you don't need to. Download a program and create a new wallet. You can create multiple wallets at no cost.</b></span>
+                </p>
+                <p>
+                  <span>Q. Do I need to provide contact information such as an email address or phone numbers to make a wallet?</span><br>
+                  <span><b>A. No, you don't need to.</b></span>
+                </p>
+                <p>
+                  <span>Q. Is there any cost for keeping a wallet?</span><br>
+                  <span><b>A. No. Keeping a wallet are free of charge.</b></span>
+                </p>
+                <p>
+                  <span>Q. Will VIPS be listed on a cryptcurrency exchange?</span><br>
+                  <span><b>A. As of April 2018, VIPS is not listed on any exchanges and does not have any value at this point. We are targeting to have VIPS listed on an exchange shortly. </b></span>
+                </p>
+                <p>
+                  <span>Q. How can I get some VIPS coins before listed. </span><br>
+                  <span><b>A. We had terminated airdrops, so there's no way to get VIPS before listed at this point. </b></span>
+                </p>
+                <p>
+                  <span>Q. What is the plan for increasing the value of VIPS?</span><br>
+                  <span><b>A. We plan to increase the number of people who have heard of VIPS at first; then we release an app that enables holders to use VIPS as a payment method. Detailed plans are to be developed.</b></span>
+                </p>
+                <p>
+                  <span>Q. What is NUQUMORITY?</span><br>
+                  <span><b>A. NUQUMORITY is the word originated in the VIP community. The word is the combination of a Japanese word 'Nukumori'(tenderness, warmness) and an English word 'Quality.' In 5ch (formerly known as 2ch), the word 'Quality' is sometimes used in the context of expressing an 'identity' of a board. 'NUQUMORITY' is the word to express the tenderness in the VIP community and a slogan to spread it outside the VIP community.</b></span>
+                </p>
+                <p>
+                  <span>Q. Where can I get more information on VIPS?</span><br>
+                  <span><b>A. Join our discord community to get the up-to-date information.</b></span>
+                </p>
+                </div>
+                <div class="modal-footer">
+                  <button type="button" class="button button-bw" data-dismiss="modal">CLOSE</button>
+                </div>
+              </div>
+            </div>
+          </div>
+          <!-- Modal -->
+          <div class="modal fade" id="pointModal" tabindex="-1" role="dialog" aria-labelledby="pointModalLabel">
+            <div class="modal-dialog" role="document">
+              <div class="modal-content">
+                <div class="modal-header">
+                  <button type="button" class="close" data-dismiss="modal" aria-label="Close"><span aria-hidden="true">×</span></button>
+                  <h4 class="modal-title" id="pointModalLabel">6 unique features of VIPS</h4>
+                </div>
+                <div class="modal-body">
+                <p class="modal-text">
+                  1. The maximum supply of VIPS is 70 billion. This number is relatively larger than the other cryptcurrencies (BTC:21 million, Litecoin: 84 million and Monacoin: 105 million), resulting in a more active use of sending and receiving among vippers as a means of communication. (This activity is called "Nageai" in Japanese). Moreover,<br>
+                  A larger maximum supply is favorable as a coin that can be a matrix of derivative tokens (We are targeting to be the No.1 cryptcurrency originated in Japan, which is equipped with smart contracts function). <br>
+                  In case of a stagnated price, we can burn a part of the total amount of coins and promote an increase in price and enhance the liquidity.<br>
+                  The maximum supply will be reached in 2120. (53,611,200 blocks)<br>
+                </p>
+                <p class="modal-text">
+                  2. We determine and ensure the budget needed for promoting VIPS before listed. This enables us to continue to conduct marketing, advertisement, and development for a while in the future. <br>
+                </p>
+                <p class="modal-text">
+                  3. We also keep a certain portion of VIPS for distribution ('NUQUMORITY' activities) and provide vipper with VIPS coins periodically. This will contribute to gain support from our original community continuously.<br>
+                </p>
+                <p class="modal-text">
+                  4. Since VIPS is originated from 5ch (formerly known as 2ch) as Monacoin is, strong support from the community can be expected.<br>
+                </p>
+                <p class="modal-text">
+                  5. VIPS has a clear roadmap to the goal. We will continuously incorporate new targets and update the roadmap so that people can invest their money under a clear goal.<br>
+                </p>
+                <p class="modal-text">
+                  6. VIPS has the smart contracts function, which enables a person to create a new VIPS token. In sending tokens, a sender needs VIPS as a fee so there will be the strong continuous demand of VIPS.<br>
+                </p>
+                </div>
+                <div class="modal-footer">
+                  <button type="button" class="button button-bw" data-dismiss="modal">CLOSE</button>
+                </div>
+              </div>
+            </div>
+          </div>
+        </div>
+      </div>
+    </div>
+  </div>
+</div>
