+++ en/holding.php
@@ -1,0 +1,63 @@
+<!-- Holding Section -->
+<div id="holding" class="text-center">
+  <div class="container">
+    <div class="section-title center">
+      <h2>SUPPLY</h2>
+      <hr>
+    </div>
+    <div class="row">
+      <div class="holding-header">
+      <p>The initial supply of VIPS is 60 billion.</p>
+      <p>Additional supply through PoS and PoW will be 10 billion in total.<br>（This will end in 2120.）</p>
+      </div>
+      <div class="holding-chart"><div class="chartjs-size-monitor" style="position: absolute; left: 0px; top: 0px; right: 0px; bottom: 0px; overflow: hidden; pointer-events: none; visibility: hidden; z-index: -1;"><div class="chartjs-size-monitor-expand" style="position:absolute;left:0;top:0;right:0;bottom:0;overflow:hidden;pointer-events:none;visibility:hidden;z-index:-1;"><div style="position:absolute;width:1000000px;height:1000000px;left:0;top:0"></div></div><div class="chartjs-size-monitor-shrink" style="position:absolute;left:0;top:0;right:0;bottom:0;overflow:hidden;pointer-events:none;visibility:hidden;z-index:-1;"><div style="position:absolute;width:200%;height:200%;left:0; top:0"></div></div></div>
+        <canvas id="holdingChart" width="400" height="400" style="display: block; height: 200px; width: 200px;" class="chartjs-render-monitor"></canvas>
+        <p class="holding-total">60<br>BILLIONS</p>
+      </div>
+      <input type="hidden" class="js_ctx-label-1" value="Reserve for sale">
+      <input type="hidden" class="js_ctx-label-2" value="Technological Development fee">
+      <input type="hidden" class="js_ctx-label-3" value="Marketing fee">
+      <input type="hidden" class="js_ctx-label-4" value="Airdrops">
+      <input type="hidden" class="js_ctx-label-5" value="Reserve by the dev team">
+      <div class="col-md-5 col-sm-5 col-md-offset-5 holding-text">
+      <p><span class="holding-yellow">■</span> Reserve for sale: 33% </p>
+      <p><span class="holding-orange">■</span> Technological Development fee：20%</p>
+      <p><span class="holding-red">■</span> Marketing fee: 20%</p>
+      <p><span class="holding-purple">■</span> Airdrops：16%</p>
+      <p><span class="holding-blue">■</span> Reserve by the dev team：10%</p>
+      </div>
+      <div class="col-md-12 col-sm-12">
+        <p class="holding-attention">
+          <b>Initial sale</b>
+        </p>
+        <p>
+        20 billion coins<br>
+        We sell the most part at 2 sat.<br>
+        Excess supply will be burned and eliminated.
+        </p>
+        <p class="holding-attention">
+          <b>Compensation of MINING</b>
+        </p>
+        <p>
+          For 1~2,000 blocks, the reward of PoS・PoW will be suppressed to 1VIPS.<br>
+        </p>
+        <p>
+          PoS<br>
+          VIPS holders are randomly selected as a reward winner of PoS.<br>
+          The reward rate is as follows.<br>
+          2,001～28,000 block：3,000vips<br>
+          28,001～525,600 block：9,500vips<br>
+          525,601 block：4,750vips<br>
+          From 525,601 block, the reward will be halved at each time 525,600 blocks have been mined.<br>
+          （the lower limit is 100 VIPS）<br>
+          The final block is 53,611,200th.
+        </p>
+        <p>
+          PoW<br>
+          2,001～53,611,200 block: 100vips
+        </p>
+        <p>Expected annual return: 5%</p>
+      </div>
+    </div>
+  </div>
+</div>
