+++ en/roadmap.php
@@ -1,0 +1,102 @@
+<!-- Roadmap Section -->
+<div id="roadmap" class="text-center">
+  <div class="container">
+    <div class="section-title center">
+      <h2>ROADMAP</h2>
+      <hr>
+    </div>
+    <p class="roadmap-desc">
+      We are targeting ...<br>
+      to enable VIPS holders to make easy payments with a QR code.<br>
+      to support individuals and companies that try to create an original VIPS compliant token.<br>
+      to encourage the recovery of devastated areas around the world with the spirit of 'Worldwide Nuqumority'(tenderness for the world).<br>
+      Our roadmap is shown below.
+    </p>
+    <div class="row">
+      <div class="col-lg-12">
+        <ul class="timeline">
+          <li>
+            <div class="timeline-image">
+              <h4>2018<br>FEB</h4>
+            </div>
+            <div class="timeline-panel" style="display: block;">
+              <div class="timeline-heading">
+                <h4>VIPS PROJECT STARTED</h4>
+              </div>
+              <div class="timeline-body">
+              <p>SPECIAL COMMUNITY EVENTS</p>
+              <p>AIRDROPS OF PROTOTYPE OF VIPS TOKENS (ERC20 TOKEN)</p>
+              </div>
+            </div>
+          </li>
+          <li class="timeline-inverted">
+            <div class="timeline-image">
+              <h4>2018<br>SPRING</h4>
+            </div>
+            <div class="timeline-panel" style="display: block;">
+              <div class="timeline-heading">
+                <h4>DEVELOPMENT OF AN ORIGINAL COIN</h4>
+              </div>
+              <div class="timeline-body">
+              <p>COMPLETION OF THE DEVELOPMENT OF ORIGINAL CRYPTCURRENCY</p>
+              <p>SWAPPING ETHEREUM VIPSTAR TOKENS FOR THE ORIGINAL COINS</p>
+              <p>BEING LISTED ON AN EXCHANGE</p>
+              <p>MARKETING OF VIPS / FUNDING FOR FURTHER DEVELOPMENT</p>
+              </div>
+            </div>
+          </li>
+          <li>
+            <div class="timeline-image">
+              <h4>2018<br>WINTER</h4>
+            </div>
+            <div class="timeline-panel" style="display: block;">
+              <div class="timeline-heading">
+                <h4>SERVICES DEVELOPMENT</h4>
+              </div>
+              <div class="timeline-body">
+              <p>STARTING THE DEVELOPMENT OF A SOCIAL GAMING APP* FOR VIPS HOLDERS</p>
+              <p>LAUNCHING A FLEA MARKET SERVICE FOR VIPS HOLDERS</p>
+              <p>ACQUIRING OFFICIAL RECOGNITION FROM 5CH</p>
+              <p>STARTING THE DEVELOPMENT OF A FILE UPLOADER FOR VIPS HOLDERS</p>
+              <p>STARTING THE DEVELOPMENT OF A 'WORLDWIDE NUQUMORITY' APP</p>
+              <p>*The service will be provided as a smartphone app</p>
+              </div>
+            </div>
+          </li>
+          <li class="timeline-inverted">
+            <div class="timeline-image">
+              <h4>2019</h4>
+            </div>
+            <div class="timeline-panel" style="display: block;">
+              <div class="timeline-heading">
+                <h4>SERVICES EXPANSION</h4>
+              </div>
+              <div class="timeline-body">
+              <p>DONATION TO 2020 TOKYO OLYMPIC COMMITTEE IN VIPS</p>
+              <p>DONATION TO DISASTER-STRICKEN AREAS OVER THE WORLD IN VIPS</p>
+              <p>RELEASE OF THE FILE UPLOADER SERVICE</p>
+              <p>RELEASE OF THE SOCIAL GAMING APP</p>
+              <p>COMPLETION OF BETA VERSION OF THE 'WORLDWIDE NUQUMORITY' APP</p>
+              </div>
+            </div>
+          </li>
+          <li>
+            <div class="timeline-image">
+              <h4>2020~</h4>
+            </div>
+            <div class="timeline-panel" style="display: block;">
+              <div class="timeline-heading">
+                <h4>LAUNCHING THE 'WORLDWIDE NUQUMORITY' SERVICE</h4>
+              </div>
+              <div class="timeline-body">
+              <p>The Worldwide Nuqumority is a service that shows organizations in the disaster-stricken areas that hold VIPS wallets.</p>
+              <p>VIPS holders can donate to them easily by just pushing a button one time</p>
+              <p>VIPS will reach the receiver's wallet faster than sending fiat money and could make purchases without withdrawing money from banks where a VIPS settlement system has been developed.
+              </div>
+            </div>
+          </li>
+        </ul>
+      </div>
+    </div>
+  </div>
+</div>
