+++ index.php
@@ -1,688 +1,25 @@
 <?php require_once('./controller.php'); ?>
 <!DOCTYPE html>
 <html lang="ja">
-<head>
-<!-- Global site tag (gtag.js) - Google Analytics -->
-<script async src="https://www.googletagmanager.com/gtag/js?id=UA-117099829-1"></script>
-<script>
-  window.dataLayer = window.dataLayer || [];
-  function gtag(){dataLayer.push(arguments);}
-  gtag('js', new Date());
-
-  gtag('config', 'UA-117099829-1');
-</script>
-
-<meta charset="utf-8">
-<meta name="viewport" content="width=device-width, initial-scale=1">
-<title>VIPSTARCOIN</title>
-<meta name="description"
-      content="VIPSTARCOIN(VIPS)は5ちゃんねるニュー速VIP板発祥の仮想通貨です。国産仮想通貨として多目的に利用されること、ひいては寄付への利用で世界に貢献することを目指して開発が進んでいます。強いコミュニティの力で運営し、盛り上げていきます。">
-<meta name="keywords" content="VIPSTAR,VIPSTARCOIN,VIPS,VIP,ニュー速VIP,仮想通貨">
-
-<!-- Favicons
-    ================================================== -->
-<link rel="shortcut icon" href="<?php echo URL; ?>/img/favicon.png" type="image/x-icon">
-<link rel="apple-touch-icon" href="<?php echo URL; ?>/img/apple-touch-icon.png">
-
-<meta property="og:title" content="VIPSTARCOIN">
-<meta property="og:type" content="website">
-<meta property="og:url" content="<?php echo URL; ?>/">
-<meta property="og:image" content="<?php echo URL; ?>/img/og-image.jpg">
-<meta property="og:site_name" content="VIPSTARCOIN">
-<meta property="og:description"
-      content="VIPSTARCOIN(VIPS)は5ちゃんねるニュー速VIP板発祥の仮想通貨です。国産仮想通貨として多目的に利用されること、ひいては寄付への利用で世界に貢献することを目指して開発が進んでいます。強いコミュニティの力で運営し、盛り上げていきます。">
-<meta name="twitter:card" content="summary_large_image">
-
-<!-- Bootstrap -->
-<link rel="stylesheet"
-      href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/css/bootstrap.min.css"
-      integrity="sha384-BVYiiSIFeK1dGmJRAkycuHAHRg32OmUcww7on3RYdg4Va+PmSTsz/K68vbdEjh4u"
-      crossorigin="anonymous">
-<link href="https://use.fontawesome.com/releases/v5.0.8/css/all.css" rel="stylesheet">
-<!-- Stylesheet
-    ================================================== -->
-<link rel="stylesheet" type="text/css"  href="<?php echo URL; ?>/css/style.css">
-<link rel="stylesheet" type="text/css" href="<?php echo URL; ?>/css/prettyPhoto.css">
-<link href='http://fonts.googleapis.com/css?family=Lato:400,700,900,300' rel='stylesheet' type='text/css'>
-<link href='http://fonts.googleapis.com/css?family=Open+Sans:400,700,800,600,300' rel='stylesheet' type='text/css'>
-<!-- HTML5 shim and Respond.js for IE8 support of HTML5 elements and media queries -->
-<!-- WARNING: Respond.js doesn't work if you view the page via file:// -->
-<!--[if lt IE 9]>
-  <script src="https://oss.maxcdn.com/html5shiv/3.7.2/html5shiv.min.js"></script>
-  <script src="https://oss.maxcdn.com/respond/1.4.2/respond.min.js"></script>
-<![endif]-->
-</head>
-<body id="page-top" data-spy="scroll" data-target=".navbar-fixed-top">
-<!-- Header -->
-<header id="header" class="header">
-  <img src="<?php echo URL; ?>/img/header.jpg" alt="VIPSTAR" class="header-image">
-</header>
-<!-- Navigation -->
-<div id="nav">
-  <nav class="navbar navbar-custom">
-    <div class="container">
-      <div class="navbar-header">
-        <button type="button" class="navbar-toggle" data-toggle="collapse" data-target=".navbar-main-collapse">
-          <i class="fa fa-bars"></i>
-        </button>
-        <a class="navbar-brand page-scroll" href="#page-top">VIPSTARCOIN</a>
-      </div>
-
-      <!-- Collect the nav links, forms, and other content for toggling -->
-      <div class="collapse navbar-collapse navbar-right navbar-main-collapse">
-        <ul class="nav navbar-nav">
-          <!-- Hidden li included to remove active class from about link when scrolled up past about section -->
-          <li class="hidden"> <a href="#page-top"></a> </li>
-          <li> <a class="page-scroll" href="#about">ABOUT</a> </li>
-          <li> <a class="page-scroll" href="#roadmap">ROADMAP</a> </li>
-          <li> <a class="page-scroll" href="#holding">HOLDING</a> </li>
-          <li> <a class="page-scroll" href="#wallet">WALLET</a> </li>
-          <li> <a class="page-scroll" href="#news">NEWS</a> </li>
-          <li> <a class="page-scroll" href="#community">COMMUNITY</a> </li>
-          <li> <a class="page-scroll" href="#contact">CONTACT</a> </li>
-        </ul>
-      </div>
-    </div>
-  </nav>
-</div>
-<!-- About Section -->
-<div id="about">
-  <div class="container">
-    <div class="dropdown">
-      <button class="button button-bw dropdown-toggle" type="button" id="dropdownMenu1" data-toggle="dropdown" aria-haspopup="true" aria-expanded="true">
-        <i class="fas fa-language"></i>LANGUAGE
-        <span class="caret"></span>
-      </button>
-      <ul class="dropdown-menu" aria-labelledby="dropdownMenu1">
-        <li class="selected">日本語<i class="fas fa-check"></i></li>
-        <li role="separator" class="divider"></li>
-        <li><a href="<?php echo URL; ?>/vip/">VIP</a></li>
-        <li><a href="<?php echo URL; ?>/en/">English</a></li>
-      </ul>
-    </div>
-
-    <div class="section-title text-center center">
-      <h2>ABOUT</h2>
-      <hr>
-    </div>
-    <div class="row">
-      <div class="col-md-12 text-center"><img src="<?php echo URL; ?>/img/about.jpg" class="js_about-img img-responsive"></div>
-      <div class="col-md-8 col-md-offset-2">
-        <div class="about-text">
-          <p>VIPSTARCOIN (VIPS) は5ちゃんねるニュー速VIP板発祥の仮想通貨です。</p>
-          <p>国産仮想通貨として多目的に利用されること、<br>ひいては寄付への利用で世界に貢献することを目指して開発が進んでいます。</p>
-          <p>強いコミュニティの力で運営し、盛り上げていきます。</p>
-          <p><a href="#" class="about-qa" data-toggle="modal" data-target="#qaModal">【超初心者用】仮想通貨って何？Q＆A&nbsp;&nbsp;<i class="far fa-window-restore"></i></a></p>
-          <p class="bl_about-button text-center">
-            <a class="button button-bw" href="#" data-toggle="modal" data-target="#pointModal">VIPS 6つのポイント&nbsp;&nbsp;<i class="far fa-window-restore"></i></a>
-          </p>
-          <!-- Modal -->
-          <div class="modal fade" id="qaModal" tabindex="-1" role="dialog" aria-labelledby="qaModalLabel">
-            <div class="modal-dialog" role="document">
-              <div class="modal-content">
-                <div class="modal-header">
-                  <button type="button" class="close" data-dismiss="modal" aria-label="Close"><span aria-hidden="true">&times;</span></button>
-                  <h4 class="modal-title" id="qaModalLabel">仮想通貨って何？Q＆A</h4>
-                </div>
-                <div class="modal-body">
-                  <p>
-                    <span>Q. VIPSって何？</span><br>
-                    <span><b>A. 仮想通貨です。</b></span>
-                  </p>
-                  <p>
-                    <span>Q. 仮想通貨って何？</span><br>
-                    <span><b>A. ブロックチェーンで管理されている、不正ができない電子通貨です。</b></span>
-                  </p>
-                  <p>
-                    <span>Q. 仮想通貨を払ったり貰ったりするには？</span><br>
-                    <span><b>A. ウォレット(財布)というものを作ります。</b></span>
-                  </p>
-                  <p>
-                    <span>Q. ウォレットを作るために入力する情報はなんですか？</span><br>
-                    <span><b>A. ありません。</b></span>
-                  </p>
-                  <p>
-                    <span>Q. 個人情報とか電話番号とかはいりますか？</span><br>
-                    <span><b>A. 不要です。</b></span>
-                  </p>
-                  <p>
-                    <span>Q. ウォレットの管理にいくら必要ですか？</span><br>
-                    <span><b>A. 無料です。</b></span>
-                  </p>
-                  <p>
-                    <span>Q. VIPSには価値があるの？</span><br>
-                    <span><b>A. ホルダーの数が多くなり、世間から通貨として認められれば価値が付きます。今は０円です。</b></span>
-                  </p>
-                  <p>
-                    <span>Q. VIPSがうまくいかなかった場合損しますか？</span><br>
-                    <span><b>A. 今のところVIPSはただでもらえるので、失敗しても失うのは時間だけです。</b></span>
-                  </p>
-                  <p>
-                    <span>Q. VIPSがうまくいったら儲かるの？</span><br>
-                    <span><b>A. 儲かると思いますが、うまくいかせるためにはホルダーを増やさなければなりません。</b></span>
-                  </p>
-                  <p>
-                    <span>Q. どうやったらホルダー増えるの？</span><br>
-                    <span><b>A. あなたが貰ったVIPSを、様々な人に分けてあげてください。</b></span>
-                  </p>
-                  <p>
-                    <span>Q. つまりあなたが私にVIPSをくれた理由は？</span><br>
-                    <span><b>A. VIPSのホルダー数を増やして、VIPSに価値を付けたいからです。共にがんばりましょう。</b></span>
-                  </p>
-                </div>
-                <div class="modal-footer">
-                  <button type="button" class="button button-bw" data-dismiss="modal">CLOSE</button>
-                </div>
-              </div>
-            </div>
-          </div>
-          <!-- Modal -->
-          <div class="modal fade" id="pointModal" tabindex="-1" role="dialog" aria-labelledby="pointModalLabel">
-            <div class="modal-dialog" role="document">
-              <div class="modal-content">
-                <div class="modal-header">
-                  <button type="button" class="close" data-dismiss="modal" aria-label="Close"><span aria-hidden="true">&times;</span></button>
-                  <h4 class="modal-title" id="pointModalLabel">VIPS 6つのポイント</h4>
-                </div>
-                <div class="modal-body">
-                  <p class="modal-text">
-                    1. VIPSの発行枚数上限は700億枚と仮想通貨の中でも比較的多く設定されているため、投げ合いやすい環境が長く続く可能性が高い。<br>
-                    他仮想通貨の発行枚数上限例）ビットコイン：2100万枚　ライトコイン：8400万枚　モナコイン：1億512万枚<br>
-                    また、他に以下の長所がある。<br>
-                    ・VIPSはスマートコントラクトで日本一を狙っているため、枚数が多いほうがトークンの母体としての利便性に長ける。<br>
-                    ・バーンを材料として取り入れていく予定なので相場が停滞したらバーンすることで価格向上・流動の促進を狙うことができる。<br>
-                    また、最大発行の700億枚には2120年に到達予定である（53,611,200ブロック）。<br>
-                  </p>
-                  <p class="modal-text">
-                    2. VIPSは広報・開発用保持がしっかりあるため、自らの地位向上のための広報活動・営業活動・開発に予算をかけることができる。<br>
-                  </p>
-                  <p class="modal-text">
-                    3. VIPSはﾇｸﾓﾘﾃｨ(配布)用の保持が多くあるのでVIPPERを中心に恒常的に配布を与えることができる。<br>
-                    これにより流動性を保つことができる。(ただし相場を破壊する程の多額の配布は実施しない)
-                  </p>
-                  <p class="modal-text">
-                    4. VIPSはモナコインと同じくコミュニティを地盤とした独自通貨であるため、コミュニティの力による発展が強く望める。
-                  </p>
-                  <p class="modal-text">
-                    5. VIPSはロードマップが定められている。そしてそこに定期的に新たな目標を組み込み進化していく。そのためロードマップが陳腐化せず段階的な価格向上が期待できる。
-                  </p>
-                  <p class="modal-text">
-                    6. VIPSはスマートコントラクト搭載通貨のため、VIPSのトークンを生み出すことができ、例えば誰かがギココイン（VIPSトークン）などを発行した場合、ギココインホルダーもVIPSを入手する必要が出てくる。
-                  </p>
-                </div>
-                <div class="modal-footer">
-                  <button type="button" class="button button-bw" data-dismiss="modal">CLOSE</button>
-                </div>
-              </div>
-            </div>
-          </div>
-        </div>
-      </div>
-    </div>
-  </div>
-</div>
-<!-- Roadmap Section -->
-<div id="roadmap" class="text-center">
-  <div class="container">
-    <div class="section-title center">
-      <h2>ROADMAP</h2>
-      <hr>
-    </div>
-    <p class="roadmap-desc">
-      URIを活用し全国のお店でQRコードを活用して簡単にお買い物ができるようになることや、<br>
-      個人または企業が様々なトークンをVIPSから生み出せるようになることは当然として、<br>
-      ワールドワイドヌクモリティという目的もあります！<br>
-      以下がそのロードマップです！
-    </p>
-    <div class="row">
-      <div class="col-lg-12">
-        <ul class="timeline">
-          <li>
-            <div class="timeline-image">
-              <h4>2018<br>FEB</h4>
-            </div>
-            <div class="timeline-panel">
-              <div class="timeline-heading">
-                <h4>VIPS発行</h4>
-              </div>
-              <div class="timeline-body">
-                <p>各種イベント</p>
-                <p>VIPを中心にAirDrop</p>
-              </div>
-            </div>
-          </li>
-          <li class="timeline-inverted">
-            <div class="timeline-image">
-              <h4>2018<br>SPRING</h4>
-            </div>
-            <div class="timeline-panel">
-              <div class="timeline-heading">
-                <h4>独自通貨開発開始</h4>
-              </div>
-              <div class="timeline-body">
-                <p>独自通貨完成＆イーサリアムトークンからのスワップ</p>
-                <p>海外取引所上場</p>
-                <p>VIPSの周知・資金調達</p>
-              </div>
-            </div>
-          </li>
-          <li>
-            <div class="timeline-image">
-              <h4>2018<br>WINTER</h4>
-            </div>
-            <div class="timeline-panel">
-              <div class="timeline-heading">
-                <h4>サービス開発開始</h4>
-              </div>
-              <div class="timeline-body">
-                <p>VIPSを使用したソーシャルゲームアプリケーションの委託開発開始</p>
-                <p>VIPSを利用したフリーマーケットサービスローンチ</p>
-                <p>ジム（5ちゃんねる管理人）と商談し正式な5chコインを目指す</p>
-                <p>VIPSを払わないと見れない仕組みのアップローダー開発開始</p>
-                <p>ワールドワイドヌクモリティ計画のためのアプリ制作</p>
-              </div>
-            </div>
-          </li>
-          <li class="timeline-inverted">
-            <div class="timeline-image">
-              <h4>2019</h4>
-            </div>
-            <div class="timeline-panel">
-              <div class="timeline-heading">
-                <h4>サービス拡大</h4>
-              </div>
-              <div class="timeline-body">
-                <p>VIPSでオリンピック寄付</p>
-                <p>VIPSで全世界の被災に寄付</p>
-                <p>アップローダー完成サービスローンチ</p>
-                <p>ソーシャルゲーム完成サービスローンチ</p>
-                <p>ワールドワイドヌクモリティアプリβ版開発完了</p>
-              </div>
-            </div>
-          </li>
-          <li>
-            <div class="timeline-image">
-              <h4>2020~</h4>
-            </div>
-            <div class="timeline-panel">
-              <div class="timeline-heading">
-                <h4>ワールドワイドヌクモリティローンチ</h4>
-              </div>
-              <div class="timeline-body">
-                <p>ワールドワイドヌクモリティとは</p>
-                <p>世界の被災情報を自動で取得して寄付すべき場所(VIPSウォレットを持つ団体)を提示してくれるシステムです</p>
-                <p>ワンプッシュでかんたんに支援することができます</p>
-                <p>法定通貨だと送金時間などの問題が発生しますが、VIPSならすぐ届き、VIPSで買い物をできるインフラが整っていれば十分な支援が望めます</p>
-                </p>
-              </div>
-            </div>
-          </li>
-        </ul>
-      </div>
-    </div>
-  </div>
-</div>
-<!-- Holding Section -->
-<div id="holding" class="text-center">
-  <div class="container">
-    <div class="section-title center">
-      <h2>HOLDING</h2>
-      <hr>
-    </div>
-    <div class="row">
-      <div class="holding-header">
-        <p>発行数600億枚</p>
-        <p>PoS・PoW合計100億枚<br>（2120年に最大発行完了予定）</p>
-      </div>
-      <div class="holding-chart">
-        <canvas id="holdingChart" width="200" height="200"></canvas>
-        <p class="holding-total">発行数<br>600億枚</p>
-      </div>
-      <div class="col-md-5 col-sm-5 col-md-offset-5 holding-text">
-        <p><span class="holding-yellow">■</span> 上場時販売：33%</p>
-        <p><span class="holding-orange">■</span> 開発：20%</p>
-        <p><span class="holding-red">■</span> マーケティング：20%</p>
-        <p><span class="holding-purple">■</span> 各種媒体での配布：16%</p>
-        <p><span class="holding-blue">■</span> 運営：10%</p>
-      </div>
-      <input type="hidden" class="js_ctx-label-1" value="上場時販売">
-      <input type="hidden" class="js_ctx-label-2" value="開発">
-      <input type="hidden" class="js_ctx-label-3" value="マーケティング">
-      <input type="hidden" class="js_ctx-label-4" value="各種媒体での配布">
-      <input type="hidden" class="js_ctx-label-5" value="運営">
-      <div class="col-md-12 col-sm-12">
-        <p>（小数点以下略）</p>
-        <p class="holding-attention">
-          <b>上場時販売</b>
-        </p>
-        <p>
-          200億枚<br>
-          大半を2satで売却<br>
-          売れ残りはバーン
-        </p>
-        <p class="holding-attention">
-          <b>マイニング報酬</b>
-        </p>
-        <p>
-          2000ブロック目までPoS・PoW、それぞれ報酬1vips<br>
-          （公開前に運営がマイニング報酬を貰わないようにするため）
-        </p>
-        <p>
-          PoS<br>
-          PoS ver3のため抽選式PoS。当選するとPoS報酬が貰える。<br>
-          2,001～28,000ブロック：PoS 3,000vips<br>
-          28,001～525,600ブロック：PoS 9,500vips<br>
-          525,601ブロック：PoS 4,750vips<br>
-          以降525,600ブロックごとに半減（下限100vips）<br>
-          53,611,200ブロックで配布完了予定
-        </p>
-        <p>
-          PoW<br>
-          2,001～53,611,200ブロック：PoW 100vips
-        </p>
-        <p>年利期待値約5%</p>
-      </div>
-    </div>
-  </div>
-</div>
-<!-- Wallet Section -->
-<div id="wallet">
-  <div class="container">
-    <div class="section-title text-center center">
-      <h2>WALLET</h2>
-      <hr>
-    </div>
-    <div class="row">
-      <div class="col-md-8 col-md-offset-2">
-        <div class="wallet-swap">
-          <p class="wallet-swap-title-b">旧VIPS(イーサリアムトークン)を独自通貨にスワップします。ご期待ください。</p>
-          <button type="button" class="button button-bw" data-toggle="modal" data-target="#walletModal">
-            スワップの約束&nbsp;&nbsp;<i class="far fa-window-restore"></i>
-          </button>
-          <!-- Modal -->
-          <div class="modal fade" id="walletModal" tabindex="-1" role="dialog" aria-labelledby="walletModalLabel">
-            <div class="modal-dialog" role="document">
-              <div class="modal-content">
-                <div class="modal-header">
-                  <button type="button" class="close" data-dismiss="modal" aria-label="Close"><span aria-hidden="true">&times;</span></button>
-                  <h4 class="modal-title" id="walletModalLabel">スワップの約束</h4>
-                </div>
-                <div class="modal-body">
-                  <p class="modal-text">
-                    <span>1. 500億枚から600億枚に発行数が増えていますが？</span><br>
-                    <span><b>1.2倍に増量してスワップを行います。</b></span>
-                  </p>
-                  <p class="modal-text">
-                    <span>2. 端数はどうなりますか？</span><br>
-                    <span><b>端数は繰り上げます。</b></span>
-                  </p>
-                  <p class="modal-text">
-                    <span>3. スワップされた分はPoSを貰えないなど制限はありますか？</span><br>
-                    <span><b>いいえ。PoSも貰うことができます。</b></span>
-                  </p>
-                  <p class="modal-text">
-                    <span>4. スマートフォンしか持っていませんが、スワップを受け取ることはできますか？</span><br>
-                    <span><b>CoinExchange上場後に送金する特別措置を用意します。ただし、行うのは上場後しばらく経ってからになり、1.2倍の増量も行わないため、パソコンを用意することをおすすめします。</b></span>
-                  </p>
-                </div>
-                <div class="modal-footer">
-                  <button type="button" class="button button-bw" data-dismiss="modal">CLOSE</button>
-                </div>
-              </div>
-            </div>
-          </div>
-        </div>
-      </div>
-    </div>
-  </div>
-</div>
-<!-- News Section -->
-<div id="news">
-  <div class="container">
-    <div class="section-title text-center center">
-      <h2>NEWS</h2>
-      <hr>
-    </div>
-    <div class="categories">
-      <ul class="cat">
-        <li>
-          <ol class="type">
-            <li><a href="#" data-filter="*">すべて</a></li>
-            <li><a href="#" data-filter=".new" class="active">最新</a></li>
-            <li><a href="#" data-filter=".develop">開発情報</a></li>
-            <li><a href="#" data-filter=".event">イベント</a></li>
-            <li><a href="#" data-filter=".other">その他</a></li>
-          </ol>
-        </li>
-      </ul>
-      <div class="clearfix"></div>
-    </div>
-    <div class="row">
-      <div class="news-items">
-        <div class="col-sm-6 col-md-3 col-lg-3 other new">
-          <div class="news-item">
-            <div class="hover-bg">
-              <a href="https://www.youtube.com/watch?v=iUzFW0sZ3s0" target="_blank">
-                <div class="hover-text">
-                  <h4>VIPSガールズアニメ化決定！</h4>
-                  <small>2018/04/01<br>（エイプリルフール）</small>
-                </div>
-                <div class="news-image">
-                  <img src="<?php echo URL; ?>/img/news/other-april-fool-2018.jpg" alt="VIPSガールズアニメ化決定！">
-                </div>
-              </a>
-            </div>
-          </div>
-        </div>
-        <div class="col-sm-6 col-md-3 col-lg-3 other new">
-          <div class="news-item">
-            <div class="hover-bg">
-              <a href="http://seesaawiki.jp/vips/d/%b8%f8%bc%b0%a5%ad%a5%e3%a5%e9%a5%af%a5%bf%a1%bc%a4%cb%a4%c4%a4%a4%a4%c6" target="_blank">
-                <div class="hover-text">
-                  <h4>公式キャラクター設定画公開</h4>
-                  <small>2018/03/26</small>
-                </div>
-                <div class="news-image">
-                  <img src="<?php echo URL; ?>/img/news/other-chara-sheet.png" alt="公式キャラクター設定画公開">
-                </div>
-              </a>
-            </div>
-          </div>
-        </div>
-        <div class="col-sm-6 col-md-3 col-lg-3 develop new">
-          <div class="news-item">
-            <div class="hover-bg">
-              <a href="http://seesaawiki.jp/vips/d/%c6%c8%bc%ab%c4%cc%b2%df%c6%b0%ba%ee%a5%c6%a5%b9%a5%c8" target="_blank">
-                <div class="hover-text">
-                  <h4>独自通貨動作テスト (SIKEN)</h4>
-                  <small>2018/03/20</small>
-                </div>
-                <div class="news-image">
-                  <img src="<?php echo URL; ?>/img/news/develop-skn.png" alt="独自通貨動作テスト (SIKEN)">
-                </div>
-              </a>
-            </div>
-          </div>
-        </div>
-        <div class="col-sm-6 col-md-3 col-lg-3 event new">
-          <div class="news-item">
-            <div class="hover-bg">
-              <a href="http://seesaawiki.jp/vips/d/%b8%f8%bc%b0%a5%ad%a5%e3%a5%e9%a5%af%a5%bf%a1%bc%a5%b3%a5%f3%a5%c6%a5%b9%a5%c8%b7%eb%b2%cc%c8%af%c9%bd" target="_blank">
-                <div class="hover-text">
-                  <h4>公式キャラクターコンテスト 結果発表</h4>
-                  <small>2018/03/18</small>
-                </div>
-                <div class="news-image">
-                  <img src="<?php echo URL; ?>/img/news/event-chara-result.png" alt="公式キャラクターコンテスト結果発表">
-                </div>
-              </a>
-            </div>
-          </div>
-        </div>
-        <div class="col-sm-6 col-md-3 col-lg-3 other new">
-          <div class="news-item">
-            <div class="hover-bg">
-              <a href="http://seesaawiki.jp/ryugoroshi/" target="_blank">
-                <div class="hover-text">
-                  <h4>提携ゲーム「竜殺しの物語」リリース</h4>
-                  <small>2018/03/14</small>
-                </div>
-                <div class="news-image">
-                  <img src="<?php echo URL; ?>/img/news/other-ryugoroshi.jpg" alt="提携ゲーム「竜殺しの物語」リリース">
-                </div>
-              </a>
-            </div>
-          </div>
-        </div>
-        <div class="col-sm-6 col-md-3 col-lg-3 develop new">
-          <div class="news-item">
-            <div class="hover-bg">
-              <a href="http://seesaawiki.jp/vips/d/%a1%da%bd%aa%ce%bb%a1%db%bb%ee%b8%b3%cd%d1%c6%c8%bc%ab%c4%cc%b2%dfTTT%b4%b0%c0%ae" target="_blank">
-                <div class="hover-text">
-                  <h4>独自通貨動作テスト (TESTSTAR)</h4>
-                  <small>2018/03/09</small>
-                </div>
-                <div class="news-image">
-                  <img src="<?php echo URL; ?>/img/news/develop-ttt.png" alt="独自通貨動作テスト (TESTSTAR)">
-                </div>
-              </a>
-            </div>
-          </div>
-        </div>
-        <div class="col-sm-6 col-md-3 col-lg-3 event new">
-          <div class="news-item">
-            <div class="hover-bg">
-              <a href="http://seesaawiki.jp/vips/d/VIP%a1%f9STAR%20%b8%f8%bc%b0%a5%ad%a5%e3%a5%e9%a5%af%a5%bf%a1%bc%a5%b3%a5%f3%a5%c6%a5%b9%a5%c8%20%a5%a8%a5%f3%a5%c8%a5%ea%a1%bc%b0%ec%cd%f7" target="_blank">
-                <div class="hover-text">
-                  <h4>VIP☆STAR 公式キャラクターコンテスト</h4>
-                  <small>2018/02/26</small>
-                </div>
-                <div class="news-image">
-                  <img src="<?php echo URL; ?>/img/news/event-chara.jpeg" alt="VIP☆STAR 公式キャラクターコンテスト">
-                </div>
-              </a>
-            </div>
-          </div>
-        </div>
-        <div class="col-sm-6 col-md-3 col-lg-3 event new">
-          <div class="news-item">
-            <div class="hover-bg">
-              <a href="http://seesaawiki.jp/vips/d/%c2%e81%b2%f3VIPS%a5%a2%a5%a4%a5%b3%a5%f3%c5%ea%c9%bc" target="_blank">
-                <div class="hover-text">
-                  <h4>VIP☆STAR コインアイコンコンテスト</h4>
-                  <small>2018/02/19</small>
-                </div>
-                <div class="news-image">
-                  <img src="<?php echo URL; ?>/img/news/event-icon.jpg" alt="VIP☆STAR コインアイコンコンテスト">
-                </div>
-              </a>
-            </div>
-          </div>
-        </div>
-        <div class="col-sm-6 col-md-3 col-lg-3 develop new">
-          <div class="news-item">
-            <div class="hover-bg">
-              <a href="https://ethplorer.io/address/0x7c7c392ab2f9a74bb79ab9e6b1caf6bae0f0f14c" target="_blank">
-                <div class="hover-text">
-                  <h4>VIP☆STAR 発行</h4>
-                  <small>2018/02/17</small>
-                </div>
-                <div class="news-image">
-                  <img src="<?php echo URL; ?>/img/news/develop-init.png" alt="VIP☆STAR 発行">
-                </div>
-              </a>
-            </div>
-          </div>
-        </div>
-      </div>
-    </div>
-  </div>
-</div>
-<!-- Community Section -->
-<div id="community" class="text-center">
-  <div class="container">
-    <div class="section-title center">
-      <h2>COMMUNITY</h2>
-      <hr>
-    </div>
-    <div class="row col-md-8 col-md-offset-2">
-      <div class="col-md-6 col-sm-6 wow fadeInDown" data-wow-delay="400ms">
-        <a href="https://discordapp.com/invite/K5gPZ2h" target="_blank">
-          <div class="community-box">
-            <i class="fab fa-discord fa-5x"></i>
-            <span class="count">2016</span>
-            <h4>Members</h4>
-          </div>
-        </a>
-      </div>
-      <div class="col-md-6 col-sm-6 wow fadeInDown" data-wow-delay="200ms">
-        <a href="https://twitter.com/VIPSTARTOKEN" target="_blank">
-          <div class="community-box">
-            <i class="fab fa-twitter fa-5x"></i>
-            <span class="count">1882</span>
-            <h4>Followers</h4>
-          </div>
-        </a>
-      </div>
-    </div>
-  </div>
-</div>
-<!-- Contact Section -->
-<div id="contact" class="text-center">
-  <div class="container">
-    <div class="section-title center">
-      <h2>CONTACT</h2>
-      <hr>
-    </div>
-    <div class="col-md-8 col-md-offset-2">
-      <div class="col-md-6">
-        <a href="mailto:vipstarcoin&#64;gmail.com">
-          <i class="far fa-envelope fa-5x"></i>
-          <p>vipstarcoin&#64;gmail.com</p>
-        </a>
-      </div>
-      <div class="col-md-6">
-        <a href="https://twitter.com/VIPSTARTOKEN" target="_blank">
-          <i class="fab fa-twitter fa-5x"></i>
-          <p>@VIPSTARTOKEN</p>
-        </a>
-      </div>
-      <div class="clearfix"></div>
-      <div class="contact-tweet">
-        <a class="twitter-timeline"
-           data-width="500"
-           data-height="500"
-           data-theme="light"
-           href="https://twitter.com/VIPSTARTOKEN?ref_src=twsrc%5Etfw">
-          Tweets by VIPSTARTOKEN
-        </a>
-        <script async src="https://platform.twitter.com/widgets.js" charset="utf-8"></script>
-      </div>
-    </div>
-  </div>
-</div>
-<div id="footer">
-  <div class="container text-center">
-    <div class="fnav">
-      <div class="addthis_inline_share_toolbox"></div>
-      <p>Copyright &copy; 2018 VIPSTARCOIN. All Rights Reserved.</p>
-    </div>
-  </div>
-</div>
-<script src="https://code.jquery.com/jquery-2.2.4.min.js"
-        integrity="sha256-BbhdlvQf/xTY9gja0Dq3HiwQF8LaCRTXxZKRutelT44="
-        crossorigin="anonymous"></script>
-<script src="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/js/bootstrap.min.js"
-        integrity="sha384-Tc5IQib027qvyjSMfHjOMaLkfuWVxZxUPnCJA7l2mCWNIpG9mGCD8wGNIcPD7Txa"
-        crossorigin="anonymous"></script>
-<script src="https://cdnjs.cloudflare.com/ajax/libs/Chart.js/2.7.2/Chart.bundle.min.js"
-        integrity="sha256-XF29CBwU1MWLaGEnsELogU6Y6rcc5nCkhhx89nFMIDQ="
-        crossorigin="anonymous"></script>
-<script type="text/javascript" src="<?php echo URL; ?>/js/jquery.isotope.js"></script>
-<script type="text/javascript" src="<?php echo URL; ?>/js/jquery.counterup.js"></script>
-<script type="text/javascript" src="<?php echo URL; ?>/js/waypoints.js"></script>
-<script type="text/javascript" src="<?php echo URL; ?>/js/jquery.inview.min.js"></script>
-<script type="text/javascript" src="<?php echo URL; ?>/js/main.js"></script>
-<script type="text/javascript" src="//s7.addthis.com/js/300/addthis_widget.js#pubid=ra-5ab60eab9e6f635f"></script>
-</body>
+  <head>
+    <?php require_once('./common/head.php'); ?>
+    <meta name="description"
+          content="VIPSTARCOIN(VIPS)は5ちゃんねるニュー速VIP板発祥の仮想通貨です。国産仮想通貨として多目的に利用されること、ひいては寄付への利用で世界に貢献することを目指して開発が進んでいます。強いコミュニティの力で運営し、盛り上げていきます。">
+    <meta name="keywords" content="VIPSTAR,VIPSTARCOIN,VIPS,VIP,ニュー速VIP,仮想通貨">
+    <meta property="og:description"
+          content="VIPSTARCOIN(VIPS)は5ちゃんねるニュー速VIP板発祥の仮想通貨です。国産仮想通貨として多目的に利用されること、ひいては寄付への利用で世界に貢献することを目指して開発が進んでいます。強いコミュニティの力で運営し、盛り上げていきます。">
+  </head>
+  <body id="page-top" data-spy="scroll" data-target=".navbar-fixed-top">
+    <?php require_once('./common/header.php'); ?>
+    <?php require_once('./common/nav.php'); ?>
+    <?php require_once('./ja/about.php'); ?>
+    <?php require_once('./ja/roadmap.php'); ?>
+    <?php require_once('./ja/holding.php'); ?>
+    <?php require_once('./ja/wallet.php'); ?>
+    <?php require_once('./common/news.php'); ?>
+    <?php require_once('./common/community.php'); ?>
+    <?php require_once('./common/contact.php'); ?>
+    <?php require_once('./ja/footer.php'); ?>
+    <?php require_once('./common/script.php'); ?>
+  </body>
 </html>
