+++ ja/about.php
@@ -1,0 +1,136 @@
+<!-- About Section -->
+<div id="about">
+  <div class="container">
+    <div class="dropdown">
+      <button class="button button-bw dropdown-toggle" type="button" id="dropdownMenu1" data-toggle="dropdown" aria-haspopup="true" aria-expanded="true">
+        <i class="fas fa-language"></i>LANGUAGE
+        <span class="caret"></span>
+      </button>
+      <ul class="dropdown-menu" aria-labelledby="dropdownMenu1">
+        <li class="selected">日本語<i class="fas fa-check"></i></li>
+        <li role="separator" class="divider"></li>
+        <li><a href="<?php echo URL; ?>/vip/">VIP</a></li>
+        <li><a href="<?php echo URL; ?>/en/">English</a></li>
+      </ul>
+    </div>
+
+    <div class="section-title text-center center">
+      <h2>ABOUT</h2>
+      <hr>
+    </div>
+    <div class="row">
+      <div class="col-md-12 text-center"><img src="<?php echo URL; ?>/img/about.jpg" class="js_about-img img-responsive"></div>
+      <div class="col-md-8 col-md-offset-2">
+        <div class="about-text">
+          <p>VIPSTARCOIN (VIPS) は5ちゃんねるニュー速VIP板発祥の仮想通貨です。</p>
+          <p>国産仮想通貨として多目的に利用されること、<br>ひいては寄付への利用で世界に貢献することを目指して開発が進んでいます。</p>
+          <p>強いコミュニティの力で運営し、盛り上げていきます。</p>
+          <p><a href="#" class="about-qa" data-toggle="modal" data-target="#qaModal">【超初心者用】仮想通貨って何？Q＆A&nbsp;&nbsp;<i class="far fa-window-restore"></i></a></p>
+          <p class="bl_about-button text-center">
+            <a class="button button-bw" href="#" data-toggle="modal" data-target="#pointModal">VIPS 6つのポイント&nbsp;&nbsp;<i class="far fa-window-restore"></i></a>
+          </p>
+          <!-- Modal -->
+          <div class="modal fade" id="qaModal" tabindex="-1" role="dialog" aria-labelledby="qaModalLabel">
+            <div class="modal-dialog" role="document">
+              <div class="modal-content">
+                <div class="modal-header">
+                  <button type="button" class="close" data-dismiss="modal" aria-label="Close"><span aria-hidden="true">&times;</span></button>
+                  <h4 class="modal-title" id="qaModalLabel">仮想通貨って何？Q＆A</h4>
+                </div>
+                <div class="modal-body">
+                  <p>
+                    <span>Q. VIPSって何？</span><br>
+                    <span><b>A. 仮想通貨です。</b></span>
+                  </p>
+                  <p>
+                    <span>Q. 仮想通貨って何？</span><br>
+                    <span><b>A. ブロックチェーンで管理されている、不正ができない電子通貨です。</b></span>
+                  </p>
+                  <p>
+                    <span>Q. 仮想通貨を払ったり貰ったりするには？</span><br>
+                    <span><b>A. ウォレット(財布)というものを作ります。</b></span>
+                  </p>
+                  <p>
+                    <span>Q. ウォレットを作るために入力する情報はなんですか？</span><br>
+                    <span><b>A. ありません。</b></span>
+                  </p>
+                  <p>
+                    <span>Q. 個人情報とか電話番号とかはいりますか？</span><br>
+                    <span><b>A. 不要です。</b></span>
+                  </p>
+                  <p>
+                    <span>Q. ウォレットの管理にいくら必要ですか？</span><br>
+                    <span><b>A. 無料です。</b></span>
+                  </p>
+                  <p>
+                    <span>Q. VIPSには価値があるの？</span><br>
+                    <span><b>A. ホルダーの数が多くなり、世間から通貨として認められれば価値が付きます。今は０円です。</b></span>
+                  </p>
+                  <p>
+                    <span>Q. VIPSがうまくいかなかった場合損しますか？</span><br>
+                    <span><b>A. 今のところVIPSはただでもらえるので、失敗しても失うのは時間だけです。</b></span>
+                  </p>
+                  <p>
+                    <span>Q. VIPSがうまくいったら儲かるの？</span><br>
+                    <span><b>A. 儲かると思いますが、うまくいかせるためにはホルダーを増やさなければなりません。</b></span>
+                  </p>
+                  <p>
+                    <span>Q. どうやったらホルダー増えるの？</span><br>
+                    <span><b>A. あなたが貰ったVIPSを、様々な人に分けてあげてください。</b></span>
+                  </p>
+                  <p>
+                    <span>Q. つまりあなたが私にVIPSをくれた理由は？</span><br>
+                    <span><b>A. VIPSのホルダー数を増やして、VIPSに価値を付けたいからです。共にがんばりましょう。</b></span>
+                  </p>
+                </div>
+                <div class="modal-footer">
+                  <button type="button" class="button button-bw" data-dismiss="modal">CLOSE</button>
+                </div>
+              </div>
+            </div>
+          </div>
+          <!-- Modal -->
+          <div class="modal fade" id="pointModal" tabindex="-1" role="dialog" aria-labelledby="pointModalLabel">
+            <div class="modal-dialog" role="document">
+              <div class="modal-content">
+                <div class="modal-header">
+                  <button type="button" class="close" data-dismiss="modal" aria-label="Close"><span aria-hidden="true">&times;</span></button>
+                  <h4 class="modal-title" id="pointModalLabel">VIPS 6つのポイント</h4>
+                </div>
+                <div class="modal-body">
+                  <p class="modal-text">
+                    1. VIPSの発行枚数上限は700億枚と仮想通貨の中でも比較的多く設定されているため、投げ合いやすい環境が長く続く可能性が高い。<br>
+                    他仮想通貨の発行枚数上限例）ビットコイン：2100万枚　ライトコイン：8400万枚　モナコイン：1億512万枚<br>
+                    また、他に以下の長所がある。<br>
+                    ・VIPSはスマートコントラクトで日本一を狙っているため、枚数が多いほうがトークンの母体としての利便性に長ける。<br>
+                    ・バーンを材料として取り入れていく予定なので相場が停滞したらバーンすることで価格向上・流動の促進を狙うことができる。<br>
+                    また、最大発行の700億枚には2120年に到達予定である（53,611,200ブロック）。<br>
+                  </p>
+                  <p class="modal-text">
+                    2. VIPSは広報・開発用保持がしっかりあるため、自らの地位向上のための広報活動・営業活動・開発に予算をかけることができる。<br>
+                  </p>
+                  <p class="modal-text">
+                    3. VIPSはﾇｸﾓﾘﾃｨ(配布)用の保持が多くあるのでVIPPERを中心に恒常的に配布を与えることができる。<br>
+                    これにより流動性を保つことができる。(ただし相場を破壊する程の多額の配布は実施しない)
+                  </p>
+                  <p class="modal-text">
+                    4. VIPSはモナコインと同じくコミュニティを地盤とした独自通貨であるため、コミュニティの力による発展が強く望める。
+                  </p>
+                  <p class="modal-text">
+                    5. VIPSはロードマップが定められている。そしてそこに定期的に新たな目標を組み込み進化していく。そのためロードマップが陳腐化せず段階的な価格向上が期待できる。
+                  </p>
+                  <p class="modal-text">
+                    6. VIPSはスマートコントラクト搭載通貨のため、VIPSのトークンを生み出すことができ、例えば誰かがギココイン（VIPSトークン）などを発行した場合、ギココインホルダーもVIPSを入手する必要が出てくる。
+                  </p>
+                </div>
+                <div class="modal-footer">
+                  <button type="button" class="button button-bw" data-dismiss="modal">CLOSE</button>
+                </div>
+              </div>
+            </div>
+          </div>
+        </div>
+      </div>
+    </div>
+  </div>
+</div>
