+++ ja/holding.php
@@ -1,0 +1,63 @@
+<!-- Holding Section -->
+<div id="holding" class="text-center">
+  <div class="container">
+    <div class="section-title center">
+      <h2>HOLDING</h2>
+      <hr>
+    </div>
+    <div class="row">
+      <div class="holding-header">
+        <p>発行数600億枚</p>
+        <p>PoS・PoW合計100億枚<br>（2120年に最大発行完了予定）</p>
+      </div>
+      <div class="holding-chart">
+        <canvas id="holdingChart" width="200" height="200"></canvas>
+        <p class="holding-total">発行数<br>600億枚</p>
+      </div>
+      <div class="col-md-5 col-sm-5 col-md-offset-5 holding-text">
+        <p><span class="holding-yellow">■</span> 上場時販売：33%</p>
+        <p><span class="holding-orange">■</span> 開発：20%</p>
+        <p><span class="holding-red">■</span> マーケティング：20%</p>
+        <p><span class="holding-purple">■</span> 各種媒体での配布：16%</p>
+        <p><span class="holding-blue">■</span> 運営：10%</p>
+      </div>
+      <input type="hidden" class="js_ctx-label-1" value="上場時販売">
+      <input type="hidden" class="js_ctx-label-2" value="開発">
+      <input type="hidden" class="js_ctx-label-3" value="マーケティング">
+      <input type="hidden" class="js_ctx-label-4" value="各種媒体での配布">
+      <input type="hidden" class="js_ctx-label-5" value="運営">
+      <div class="col-md-12 col-sm-12">
+        <p>（小数点以下略）</p>
+        <p class="holding-attention">
+          <b>上場時販売</b>
+        </p>
+        <p>
+          200億枚<br>
+          大半を2satで売却<br>
+          売れ残りはバーン
+        </p>
+        <p class="holding-attention">
+          <b>マイニング報酬</b>
+        </p>
+        <p>
+          2000ブロック目までPoS・PoW、それぞれ報酬1vips<br>
+          （公開前に運営がマイニング報酬を貰わないようにするため）
+        </p>
+        <p>
+          PoS<br>
+          PoS ver3のため抽選式PoS。当選するとPoS報酬が貰える。<br>
+          2,001～28,000ブロック：PoS 3,000vips<br>
+          28,001～525,600ブロック：PoS 9,500vips<br>
+          525,601ブロック：PoS 4,750vips<br>
+          以降525,600ブロックごとに半減（下限100vips）<br>
+          53,611,200ブロックで配布完了予定
+        </p>
+        <p>
+          PoW<br>
+          2,001～53,611,200ブロック：PoW 100vips
+        </p>
+        <p>年利期待値約5%</p>
+      </div>
+    </div>
+  </div>
+</div>
