+++ vip/footer.php
@@ -1,0 +1,8 @@
+<div id="footer">
+  <div class="container text-center">
+    <div class="fnav">
+      <div class="addthis_inline_share_toolbox"></div>
+      <p>Copyright &copy; 2018 VIPSTARCOIN. All Rights Reserved.</p>
+    </div>
+  </div>
+</div>
