+++ vip/roadmap.php
@@ -1,0 +1,101 @@
+<!-- Roadmap Section -->
+<div id="roadmap" class="text-center">
+  <div class="container">
+    <div class="section-title center">
+      <h2>ROADMAP</h2>
+      <hr>
+    </div>
+    <p class="roadmap-desc">
+      URIを活用し全国のお店でQRコードを活用して簡単にお買い物ができるようになることや、<br>
+      個人または企業が様々なトークンをVIPSから生み出せるようになることは当然として、<br>
+      ワールドワイドヌクモリティという目的もあります！<br>
+      以下がそのロードマップです！
+    </p>
+    <div class="row">
+      <div class="col-lg-12">
+        <ul class="timeline">
+          <li>
+            <div class="timeline-image">
+              <h4>2018<br>FEB</h4>
+            </div>
+            <div class="timeline-panel">
+              <div class="timeline-heading">
+                <h4>VIPS発行</h4>
+              </div>
+              <div class="timeline-body">
+                <p>各種イベント</p>
+                <p>VIPを中心にAirDrop</p>
+              </div>
+            </div>
+          </li>
+          <li class="timeline-inverted">
+            <div class="timeline-image">
+              <h4>2018<br>SPRING</h4>
+            </div>
+            <div class="timeline-panel">
+              <div class="timeline-heading">
+                <h4>独自通貨開発開始</h4>
+              </div>
+              <div class="timeline-body">
+                <p>独自通貨完成＆イーサリアムトークンからのスワップ</p>
+                <p>海外取引所上場</p>
+                <p>VIPSの周知・資金調達</p>
+              </div>
+            </div>
+          </li>
+          <li>
+            <div class="timeline-image">
+              <h4>2018<br>WINTER</h4>
+            </div>
+            <div class="timeline-panel">
+              <div class="timeline-heading">
+                <h4>サービス開発開始</h4>
+              </div>
+              <div class="timeline-body">
+                <p>VIPSを使用したソーシャルゲームアプリケーションの委託開発開始</p>
+                <p>VIPSを利用したフリーマーケットサービスローンチ</p>
+                <p>ジムと商談し正式な5chコインを目指す</p>
+                <p>VIPSを払わないと見れない仕組みのうｐロダ開発開始</p>
+                <p>ワールドワイドヌクモリティ計画のためのアプリ制作</p>
+              </div>
+            </div>
+          </li>
+          <li class="timeline-inverted">
+            <div class="timeline-image">
+              <h4>2019</h4>
+            </div>
+            <div class="timeline-panel">
+              <div class="timeline-heading">
+                <h4>サービス拡大</h4>
+              </div>
+              <div class="timeline-body">
+                <p>VIPSでオリンピック寄付</p>
+                <p>VIPSで全世界の被災に寄付</p>
+                <p>うｐロダ完成サービスローンチ</p>
+                <p>ソシャゲ完成サービスローンチ</p>
+                <p>ワールドワイドヌクモリティアプリβ版開発完了</p>
+              </div>
+            </div>
+          </li>
+          <li>
+            <div class="timeline-image">
+              <h4>2020~</h4>
+            </div>
+            <div class="timeline-panel">
+              <div class="timeline-heading">
+                <h4>ワールドワイドヌクモリティローンチ</h4>
+              </div>
+              <div class="timeline-body">
+                <p>ワールドワイドヌクモリティとは</p>
+                <p>世界の被災情報を自動で取得して寄付すべき場所(VIPSウォレット持ちの団体)を提示してくれるシステムだお</p>
+                <p>ワンプッシュでヌクモリティ支援できるVIPPER大喜びのシステムおｗｗｗｗｗｗｗｗｗｗｗｗｗｗｗｗｗｗ</p>
+                <p>法定通貨だと送金時間などの問題が発生するけおVIPSならすぐ届くしVIPSでお買い物できるインフラが整っていれば十分な支援が望めるお</p>
+                </p>
+              </div>
+            </div>
+          </li>
+        </ul>
+      </div>
+    </div>
+  </div>
+</div>
