+++ vip/wallet.php
@@ -1,0 +1,51 @@
+<!-- Wallet Section -->
+<div id="wallet">
+  <div class="container">
+    <div class="section-title text-center center">
+      <h2>WALLET</h2>
+      <hr>
+    </div>
+    <div class="row">
+      <div class="col-md-8 col-md-offset-2">
+        <div class="wallet-swap">
+          <p class="wallet-swap-title-b">（　＾ω＾）旧VIPS(イーサリアムトークン)を独自通貨にスワップするお！乞うご期待お！</p>
+          <button type="button" class="button button-bw" data-toggle="modal" data-target="#walletModal">
+            スワップの約束&nbsp;&nbsp;<i class="far fa-window-restore"></i>
+          </button>
+          <!-- Modal -->
+          <div class="modal fade" id="walletModal" tabindex="-1" role="dialog" aria-labelledby="walletModalLabel">
+            <div class="modal-dialog" role="document">
+              <div class="modal-content">
+                <div class="modal-header">
+                  <button type="button" class="close" data-dismiss="modal" aria-label="Close"><span aria-hidden="true">&times;</span></button>
+                  <h4 class="modal-title" id="walletModalLabel">スワップの約束</h4>
+                </div>
+                <div class="modal-body">
+                  <p class="modal-text">
+                    <span>1. 500億枚から600億枚に発行数増えてんじゃねえか！</span><br>
+                    <span><b>（　＾ω＾）うはｗｗｗｗｗｗｗ1.2倍スワップしてやんおｗｗｗｗｗｗｗｗｗｗｗ</b></span>
+                  </p>
+                  <p class="modal-text">
+                    <span>2. 端数が出るけど・・・・</span><br>
+                    <span><b>（　＾ω＾）繰り上げるお！！</b></span>
+                  </p>
+                  <p class="modal-text">
+                    <span>3. は？1.2倍にしてくれるの？しかも端数繰り上げ？PoS貰えないとか罠ある？・・・・</span><br>
+                    <span><b>（　＾ω＾）PoS年利ちゃんと貰えるから持っとけおｗｗｗっうぇｗｗｗっうぇｗｗｗｗｗｗｗｗ</b></span>
+                  </p>
+                  <p class="modal-text">
+                    <span>4. 俺スマホしかないからスワップしてもらえないよな・・・</span><br>
+                    <span><b>（　＾ω＾）本当はダメなんだけお特別にCE上場したら送ってあげるお！上場してしばらくしてからになると思うし、1.2倍ボーナスもないからパソコン買ってくることをオヌヌメするけお！</b></span>
+                  </p>
+                </div>
+                <div class="modal-footer">
+                  <button type="button" class="button button-bw" data-dismiss="modal">CLOSE</button>
+                </div>
+              </div>
+            </div>
+          </div>
+        </div>
+      </div>
+    </div>
+  </div>
+</div>
